requirements.txt
asgiref==3.11.1
boto3
Django==6.0.2
django-cors-headers==4.9.0
django-storages
djangorestframework==3.16.1
psycopg[binary]
python-decouple==3.8
sqlparse==0.5.5
whitenoise
requests
